Fix: restort texts and constants, rename about -> about page
