+++ src/components/AboutPage/constants.js
@@ -1,0 +1,9 @@
+import photoAboutImg_1 from '../../images/about/about-1.jpg';
+import photoAboutImg_2 from '../../images/about/about-2.jpg';
+import photoAboutImg_3 from '../../images/about/about-3.jpg';
+
+export const photoAbout = [
+  photoAboutImg_1,
+  photoAboutImg_2,
+  photoAboutImg_3
+];
