+++ src/components/AboutPage/texts.js
@@ -1,0 +1,5 @@
+export const ABOUT = [
+  "<colored>Kenguru</colored> не просто про комфорт. Он про эстетику с первых дней. Про меньше, но лучше. Про функциональность и разумное потребление. Наша философия - это <colored>философия комфорта</colored>. Малыша и его мамы. Это помощь родителям в создании уютной, комфортной и безопасной атмосферы для малыша.",
+  "Мы за <colored>разумное потребление</colored> и за то, чтобы вещи служили как можно дольше. Мы за то, чтобы вещи были многофункциональными, именно поэтому у пелёнок есть открывающиеся плечи, а у конвертов - съёмные утеплители. Мы сохраняем наши принципы работы в каждом из своих изделий, чтобы вы имели возможность наслаждаться счастливыми моментами со своим малышом, а не беспокоиться о его комфорте.",
+  "Мы серьезно подходим к выбору материалов, которые используем. Первые, кто пользуется нашей продукцией -  это наши собственные дети и дети близких людей. Для наших изделий мы выбираем абсолютно <colored>безопасные и качественные материалы</colored>, которые не вызывают аллергию, при этом сохраняя ценовую доступность для нашего покупателя. Бренд – это команда единомышленников, объединенных целью создавать <colored>стильные и качественные вещи</colored> для малышей. Вместе с нами работают настоящие профессионалы, которые как и мы любят свое дело. Искренне верим, что вы это почувствуете с первой покупкой нашей продукции!"
+];
