+++ src/components/PopupAddCart/texts.js
@@ -1,0 +1,3 @@
+export const POPUP_ADD_CART_TITLE = 'Товар добавлен в корзину';
+export const POPUP_ADD_CART_KEEP_SHOPPING = 'Продолжить покупки';
+export const POPUP_ADD_CART_GO_TO_BASKET = 'В корзину';
