+++ src/components/PopupCare/constants.js
@@ -1,0 +1,16 @@
+import iconWash from '../../images/popup_care/wash.svg';
+import iconDrying from '../../images/popup_care/drying.svg';
+import iconIroning from '../../images/popup_care/ironing.svg';
+import iconDrycleaning from '../../images/popup_care/drycleaning.svg';
+import iconWhitening from '../../images/popup_care/whitening.svg';
+import iconClose from '../../images/popup_close.svg';
+export const popupClose = iconClose;
+
+export const popupCareIcons = {
+  wash: iconWash,
+  drying: iconDrying,
+  ironing: iconIroning,
+  drycleaning: iconDrycleaning,
+  whitening: iconWhitening
+};
+
