+++ src/components/PopupCare/texts.js
@@ -1,0 +1,11 @@
+export const POPUP_CARE_TITLE = 'Рекомендации по уходу';
+export const POPUP_CARE_WASH_TITLE = 'Стирка';
+export const POPUP_CARE_WASH_TEXT = 'Рекомендуемая температура стирки изделий из трикотажного хлопка - 40 градусов на малых и средних оборотах. Для трикотажных хлопковых вещей рекомендуется применять мягко действующие очищающие средства: гели для деликатных тканей, порошки для детского белья. Чтобы сохранить яркость цвета изделий, мы рекомендуем стирать изделия с изнанки.';
+export const POPUP_CARE_DRYING_TITLE = 'Сушка';
+export const POPUP_CARE_DRYING_TEXT = 'Конверты и съёмные утеплители следует сушить в горизонтальном положении, в расправленном виде. Трикотажные изделия достаточно развесить на сушилке в вертикальном положении. Использование сушильных машин крайне нежелательно, так как может привести к деформации изделия.';
+export const POPUP_CARE_IRONING_TITLE = 'Глажка';
+export const POPUP_CARE_IRONING_TEXT = 'Гладить любые наши изделия рекомендуется при температуре не более 100-110 градусов, с использованием марли. Трикотажные хлопковые изделия имеют свойство усадки после стирки. Вернуть изделиям прежний вид и размер поможет не только правильная стирка, но и правильное глажение. Отпаривать необходимо в щадящем режиме. Трикотажные изделия можно слегка растягивать при глажке – это вернёт их прежний вид и избавит от усадки. Ткани с блёстками и другими видами напыления гладить можно только с изнаночной стороны.';
+export const POPUP_CARE_DRYCLEANING_TITLE = 'Химчистка';
+export const POPUP_CARE_DRYCLEANING_TEXT = 'Химчистка как трикотажных, так и изделий из хлопка запрещена.';
+export const POPUP_CARE_WHITENING_TITLE = 'Отбеливание';
+export const POPUP_CARE_WHITENING_TEXT = 'Моющие средства, содержащие отбеливающие вещества и хлор могут испортить ткань, поэтому их применение не рекомендуется. Возможно использование пятновыводителя без хлора, но лучше всегда предварительно проверить действие выбранного средства на незаметном участке изделия.';
