+++ src/config/constants.js
@@ -13,16 +13,7 @@ import photoInstFive from '../images/instagram/inst-5.jpg';
 import photoInstSix from '../images/instagram/inst-6.jpg';
 import photoInstSeven from '../images/instagram/inst-7.jpg';
 import photoInstEight from '../images/instagram/inst-8.jpg';
-// компонент О бренде
-import photoAboutOneImg from '../images/about/about-1.jpg';
-import photoAboutTwoImg from '../images/about/about-2.jpg';
-import photoAboutThreeImg from '../images/about/about-3.jpg';
-import iconWash from '../images/popup_care/wash.svg';
-import iconDrying from '../images/popup_care/drying.svg';
-import iconIroning from '../images/popup_care/ironing.svg';
-import iconDrycleaning from '../images/popup_care/drycleaning.svg';
-import iconWhitening from '../images/popup_care/whitening.svg';
-import iconClose from '../images/popup_close.svg';
+import {useMediaQuery} from "react-responsive";
 
 export const instagramPhotoData = [
   { link: photoInstOne },
@@ -75,18 +66,3 @@ export const advantagesData = [
     description: 'Все вещи сделаны вручную с любовью небольшими партиями',
   },
 ];
-
-// компонент О бренде
-export const photoAboutOne = photoAboutOneImg;
-export const photoAboutTwo = photoAboutTwoImg;
-export const photoAboutThree = photoAboutThreeImg;
-export const popupClose = iconClose;
-
-export const popupCareIcons = {
-  wash: iconWash,
-  drying: iconDrying,
-  ironing: iconIroning,
-  drycleaning: iconDrycleaning,
-  whitening: iconWhitening
-};
-
