+++ src/config/texts.js
@@ -6,27 +6,6 @@ export const DELIVERY_TEXT = 'Доставка';
 export const CONTACTS_TEXT = 'Контакты';
 export const PHONE = '+7 925 343 34 34';
 
-// texts for cart
-export const POPUP_ADD_CART_TITLE = 'Товар добавлен в корзину';
-export const POPUP_ADD_CART_KEEP_SHOPPING = 'Продолжить покупки';
-export const POPUP_ADD_CART_GO_TO_BASKET = 'В корзину';
-
-export let ABOUT_1 = "<colored>Kenguru</colored> не просто про комфорт. Он про эстетику с первых дней. Про меньше, но лучше. Про функциональность и разумное потребление. Наша философия - это <colored>философия комфорта</colored>. Малыша и его мамы. Это помощь родителям в создании уютной, комфортной и безопасной атмосферы для малыша.";
-export let ABOUT_2 = "Мы за <colored>разумное потребление</colored> и за то, чтобы вещи служили как можно дольше. Мы за то, чтобы вещи были многофункциональными, именно поэтому у пелёнок есть открывающиеся плечи, а у конвертов - съёмные утеплители. Мы сохраняем наши принципы работы в каждом из своих изделий, чтобы вы имели возможность наслаждаться счастливыми моментами со своим малышом, а не беспокоиться о его комфорте.";
-export let ABOUT_3 = "Мы серьезно подходим к выбору материалов, которые используем. Первые, кто пользуется нашей продукцией -  это наши собственные дети и дети близких людей. Для наших изделий мы выбираем абсолютно <colored>безопасные и качественные материалы</colored>, которые не вызывают аллергию, при этом сохраняя ценовую доступность для нашего покупателя. Бренд – это команда единомышленников, объединенных целью создавать <colored>стильные и качественные вещи</colored> для малышей. Вместе с нами работают настоящие профессионалы, которые как и мы любят свое дело. Искренне верим, что вы это почувствуете с первой покупкой нашей продукции!";
-
-export const POPUP_CARE_TITLE = 'Рекомендации по уходу';
-export const POPUP_CARE_WASH_TITLE = 'Стирка';
-export const POPUP_CARE_WASH_TEXT = 'Рекомендуемая температура стирки изделий из трикотажного хлопка - 40 градусов на малых и средних оборотах. Для трикотажных хлопковых вещей рекомендуется применять мягко действующие очищающие средства: гели для деликатных тканей, порошки для детского белья. Чтобы сохранить яркость цвета изделий, мы рекомендуем стирать изделия с изнанки.';
-export const POPUP_CARE_DRYING_TITLE = 'Сушка';
-export const POPUP_CARE_DRYING_TEXT = 'Конверты и съёмные утеплители следует сушить в горизонтальном положении, в расправленном виде. Трикотажные изделия достаточно развесить на сушилке в вертикальном положении. Использование сушильных машин крайне нежелательно, так как может привести к деформации изделия.';
-export const POPUP_CARE_IRONING_TITLE = 'Глажка';
-export const POPUP_CARE_IRONING_TEXT = 'Гладить любые наши изделия рекомендуется при температуре не более 100-110 градусов, с использованием марли. Трикотажные хлопковые изделия имеют свойство усадки после стирки. Вернуть изделиям прежний вид и размер поможет не только правильная стирка, но и правильное глажение. Отпаривать необходимо в щадящем режиме. Трикотажные изделия можно слегка растягивать при глажке – это вернёт их прежний вид и избавит от усадки. Ткани с блёстками и другими видами напыления гладить можно только с изнаночной стороны.';
-export const POPUP_CARE_DRYCLEANING_TITLE = 'Химчистка';
-export const POPUP_CARE_DRYCLEANING_TEXT = 'Химчистка как трикотажных, так и изделий из хлопка запрещена.';
-export const POPUP_CARE_WHITENING_TITLE = 'Отбеливание';
-export const POPUP_CARE_WHITENING_TEXT = 'Моющие средства, содержащие отбеливающие вещества и хлор могут испортить ткань, поэтому их применение не рекомендуется. Возможно использование пятновыводителя без хлора, но лучше всегда предварительно проверить действие выбранного средства на незаметном участке изделия.';
-
 // texts for footer
 export const OFFER_TEXT = "Оферта";
 export const TERMS_OF_USE_TEXT = "Политика конфиденциальности";
